Update Distance to handle decimal

+++ Model/RunningSession.cs
@@ -6,7 +6,7 @@ namespace RunTracker.Model
     {
         public ObjectId Id { get; set; }
         public DateTime? Date { get; set; }
-        public float? Distance { get; set; }
+        public decimal? Distance { get; set; }
         public TimeSpan? Time { get; set; }
         public RunType? RunType { get; set; }
     }
